Create CategoriesService

+++ src/main/java/tc3005b224/amazonconnectinsights/repository/CategoriesRepository.java
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Repository;
 import tc3005b224.amazonconnectinsights.models_sql.Categories;
 
 @Repository//
-public interface CategoriesRepository extends CrudRepository<Categories, Integer> {
+public interface CategoriesRepository extends CrudRepository<Categories, Byte> {
 }
